# Frontend/swagger-api-spec.yaml
openapi: 3.0.3
info:
  title: HCMUT Tutor System API
  description: |
    API specification cho hệ thống Tutor của HCMUT.
    
    **Lưu ý quan trọng:**
    - Tất cả response đều phải theo format chuẩn ApiResponse
    - Các endpoint có pagination phải trả về PaginatedData
    - Authentication sử dụng Bearer Token trong header
    - Khi có lỗi, vẫn phải trả về ApiResponse với success=false
  version: 1.0.0
  contact:
    name: Frontend Team
    email: chinh.dang5504@hcmut.edu.vn
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.hcmut-tutor.edu.vn/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token từ SSO system
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
        - statusCode
        - message
        - data
      properties:
        success:
          type: boolean
          description: Trạng thái thành công/thất bại
        statusCode:
          type: integer
          description: HTTP status code
          example: 200
        message:
          type: string
          description: Thông báo cho người dùng
          example: "Success"
        data:
          type: object
          description: Dữ liệu trả về (có thể là object, array, hoặc null)
        timestamp:
          type: string
          format: date-time
          description: Thời gian server xử lý request
          example: "2025-11-23T16:26:31+07:00"
    PaginatedData:
      type: object
      required:
        - data
        - totalItems
        - currentPage
        - totalPages
        - pageSize
      properties:
        data:
          type: array
          items:
            type: object
          description: Mảng dữ liệu của trang hiện tại
        totalItems:
          type: integer
          description: Tổng số items trong database
          example: 100
        currentPage:
          type: integer
          description: Trang hiện tại (bắt đầu từ 1)
          example: 1
        totalPages:
          type: integer
          description: Tổng số trang
          example: 10
        pageSize:
          type: integer
          description: Số items mỗi trang
          example: 10
...
    TutorProfile:
      type: object
      properties:
        id:
          type: string
          example: "T001"
        fullName:
          type: string
          example: "Nguyen Minh Tri"
        email:
          type: string
          format: email
          example: "tutor@hcmut.edu.vn"
        phone:
          type: string
          example: "0901234567"
        address:
          type: string
          example: "123 ABC Street, HCMC"
        dateOfBirth:
          type: string
          format: date
          example: "1995-05-15"
        sex:
          type: string
          enum: [Male, Female, Other]
          example: "Male"
        bio:
          type: string
          example: "Experienced tutor with 5 years of teaching"
        teachingSubjects:
          type: array
          items:
            type: string
          example: ["Advanced Calculus", "Linear Algebra"]
        rating:
          type: number
          format: float
          example: 4.5
        avatar:
          type: string
          format: uri
          nullable: true
...
paths:
...
  /tutors/search:
    get:
      tags:
        - Tutor
      summary: Tìm kiếm tutor
      parameters:
        - name: q
          in: query
          required: true
          description: Từ khóa tìm kiếm (tên, môn học, khoa)
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PaginatedData'
                          - type: object
                            properties:
                              data:
                                type: array
                                items:
                                  $ref: '#/components/schemas/TutorProfile'
...